+++ src/CancellationToken.ts
@@ -4,38 +4,40 @@ import * as os from 'os';
 import * as path from 'path';
 import * as ts from 'typescript';
 
+import { FsHelper } from './FsHelper';
+
 interface CancellationTokenData {
   isCancelled: boolean;
   cancellationFileName: string;
 }
 
 export class CancellationToken {
-  isCancelled: boolean;
-  cancellationFileName: string;
-  lastCancellationCheckTime: number;
-  constructor(cancellationFileName: string, isCancelled: boolean) {
+  private isCancelled: boolean;
+  private cancellationFileName: string;
+  private lastCancellationCheckTime: number;
+  constructor(cancellationFileName?: string, isCancelled?: boolean) {
     this.isCancelled = !!isCancelled;
     this.cancellationFileName =
       cancellationFileName || crypto.randomBytes(64).toString('hex');
     this.lastCancellationCheckTime = 0;
   }
 
-  static createFromJSON(json: CancellationTokenData) {
+  public static createFromJSON(json: CancellationTokenData) {
     return new CancellationToken(json.cancellationFileName, json.isCancelled);
   }
 
-  toJSON() {
+  public toJSON() {
     return {
       cancellationFileName: this.cancellationFileName,
       isCancelled: this.isCancelled
     };
   }
 
-  getCancellationFilePath() {
+  public getCancellationFilePath() {
     return path.join(os.tmpdir(), this.cancellationFileName);
   }
 
-  isCancellationRequested() {
+  public isCancellationRequested() {
     if (this.isCancelled) {
       return true;
     }
@@ -46,25 +48,28 @@ export class CancellationToken {
     if (duration > 10) {
       // check no more than once every 10ms
       this.lastCancellationCheckTime = time;
-      this.isCancelled = fs.existsSync(this.getCancellationFilePath());
+      this.isCancelled = FsHelper.existsSync(this.getCancellationFilePath());
     }
 
     return this.isCancelled;
   }
 
-  throwIfCancellationRequested() {
+  public throwIfCancellationRequested() {
     if (this.isCancellationRequested()) {
       throw new ts.OperationCanceledException();
     }
   }
 
-  requestCancellation() {
+  public requestCancellation() {
     fs.writeFileSync(this.getCancellationFilePath(), '');
     this.isCancelled = true;
   }
 
-  cleanupCancellation() {
-    if (this.isCancelled && fs.existsSync(this.getCancellationFilePath())) {
+  public cleanupCancellation() {
+    if (
+      this.isCancelled &&
+      FsHelper.existsSync(this.getCancellationFilePath())
+    ) {
       fs.unlinkSync(this.getCancellationFilePath());
       this.isCancelled = false;
     }