+++ src/index.ts
@@ -2,15 +2,14 @@ import * as path from 'path';
 import * as process from 'process';
 import * as childProcess from 'child_process';
 import chalk, { Chalk } from 'chalk';
-import * as fs from 'fs';
+import * as micromatch from 'micromatch';
 import * as os from 'os';
 import * as webpack from 'webpack';
-import isString = require('lodash/isString');
-import isFunction = require('lodash/isFunction');
 import { CancellationToken } from './CancellationToken';
 import { NormalizedMessage } from './NormalizedMessage';
 import { createDefaultFormatter } from './formatter/defaultFormatter';
 import { createCodeframeFormatter } from './formatter/codeframeFormatter';
+import { FsHelper } from './FsHelper';
 import { Message } from './Message';
 
 import { AsyncSeriesHook, SyncHook } from 'tapable';
@@ -41,10 +40,12 @@ interface Options {
   tsconfig: string;
   compilerOptions: object;
   tslint: string | true;
+  tslintAutoFix: boolean;
   watch: string | string[];
   async: boolean;
   ignoreDiagnostics: number[];
   ignoreLints: string[];
+  reportFiles: string[];
   colors: boolean;
   logger: Logger;
   formatter: 'default' | 'codeframe' | Formatter;
@@ -64,52 +65,60 @@ interface Options {
  * Options description in README.md
  */
 class ForkTsCheckerWebpackPlugin {
-  static DEFAULT_MEMORY_LIMIT = 2048;
-  static ONE_CPU = 1;
-  static ALL_CPUS = os.cpus && os.cpus() ? os.cpus().length : 1;
-  static ONE_CPU_FREE = Math.max(1, ForkTsCheckerWebpackPlugin.ALL_CPUS - 1);
-  static TWO_CPUS_FREE = Math.max(1, ForkTsCheckerWebpackPlugin.ALL_CPUS - 2);
+  public static readonly DEFAULT_MEMORY_LIMIT = 2048;
+  public static readonly ONE_CPU = 1;
+  public static readonly ALL_CPUS = os.cpus && os.cpus() ? os.cpus().length : 1;
+  public static readonly ONE_CPU_FREE = Math.max(
+    1,
+    ForkTsCheckerWebpackPlugin.ALL_CPUS - 1
+  );
+  public static readonly TWO_CPUS_FREE = Math.max(
+    1,
+    ForkTsCheckerWebpackPlugin.ALL_CPUS - 2
+  );
 
-  options: Partial<Options>;
-  tsconfig: string;
-  compilerOptions: object;
-  tslint: string | true;
-  watch: string[];
-  ignoreDiagnostics: number[];
-  ignoreLints: string[];
-  logger: Logger;
-  silent: boolean;
-  async: boolean;
-  checkSyntacticErrors: boolean;
-  workersNumber: number;
-  memoryLimit: number;
-  useColors: boolean;
-  colors: Chalk;
-  formatter: Formatter;
+  public readonly options: Partial<Options>;
+  private tsconfig: string;
+  private compilerOptions: object;
+  private tslint?: string | true;
+  private tslintAutoFix: boolean;
+  private watch: string[];
+  private ignoreDiagnostics: number[];
+  private ignoreLints: string[];
+  private reportFiles: string[];
+  private logger: Logger;
+  private silent: boolean;
+  private async: boolean;
+  private checkSyntacticErrors: boolean;
+  private workersNumber: number;
+  private memoryLimit: number;
+  private useColors: boolean;
+  private colors: Chalk;
+  private formatter: Formatter;
 
-  tsconfigPath: string;
-  tslintPath: string;
-  watchPaths: string[];
+  private tsconfigPath?: string;
+  private tslintPath?: string;
+  private watchPaths: string[];
 
-  compiler: any;
-  started: [number, number];
-  elapsed: [number, number];
-  cancellationToken: CancellationToken;
+  private compiler: any;
+  private started?: [number, number];
+  private elapsed?: [number, number];
+  private cancellationToken?: CancellationToken;
 
-  isWatching: boolean;
-  checkDone: boolean;
-  compilationDone: boolean;
-  diagnostics: NormalizedMessage[];
-  lints: NormalizedMessage[];
+  private isWatching: boolean;
+  private checkDone: boolean;
+  private compilationDone: boolean;
+  private diagnostics: NormalizedMessage[];
+  private lints: NormalizedMessage[];
 
-  emitCallback: () => void;
-  doneCallback: () => void;
-  typescriptVersion: any;
-  tslintVersion: any;
+  private emitCallback: () => void;
+  private doneCallback: () => void;
+  private typescriptVersion: any;
+  private tslintVersion: any;
 
-  service: childProcess.ChildProcess;
+  private service?: childProcess.ChildProcess;
 
-  vue: boolean;
+  private vue: boolean;
 
   constructor(options?: Partial<Options>) {
     options = options || ({} as Options);
@@ -125,11 +134,12 @@ class ForkTsCheckerWebpackPlugin {
         ? './tslint.json'
         : options.tslint
       : undefined;
-    this.watch = isString(options.watch)
-      ? [options.watch]
-      : options.watch || [];
+    this.tslintAutoFix = options.tslintAutoFix || false;
+    this.watch =
+      typeof options.watch === 'string' ? [options.watch] : options.watch || [];
     this.ignoreDiagnostics = options.ignoreDiagnostics || [];
     this.ignoreLints = options.ignoreLints || [];
+    this.reportFiles = options.reportFiles || [];
     this.logger = options.logger || console;
     this.silent = options.silent === true; // default false
     this.async = options.async !== false; // default true
@@ -140,7 +150,7 @@ class ForkTsCheckerWebpackPlugin {
     this.useColors = options.colors !== false; // default true
     this.colors = new chalk.constructor({ enabled: this.useColors });
     this.formatter =
-      options.formatter && isFunction(options.formatter)
+      options.formatter && typeof options.formatter === 'function'
         ? options.formatter
         : ForkTsCheckerWebpackPlugin.createFormatter(
             (options.formatter as 'default' | 'codeframe') || 'default',
@@ -173,7 +183,7 @@ class ForkTsCheckerWebpackPlugin {
     this.vue = options.vue === true; // default false
   }
 
-  static createFormatter(type: 'default' | 'codeframe', options: any) {
+  private static createFormatter(type: 'default' | 'codeframe', options: any) {
     switch (type) {
       case 'default':
         return createDefaultFormatter();
@@ -186,18 +196,18 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  apply(compiler: webpack.Compiler) {
+  public apply(compiler: webpack.Compiler) {
     this.compiler = compiler;
 
     this.tsconfigPath = this.computeContextPath(this.tsconfig);
     this.tslintPath = this.tslint
       ? this.computeContextPath(this.tslint as string)
-      : null;
+      : undefined;
     this.watchPaths = this.watch.map(this.computeContextPath.bind(this));
 
     // validate config
-    const tsconfigOk = fs.existsSync(this.tsconfigPath);
-    const tslintOk = !this.tslintPath || fs.existsSync(this.tslintPath);
+    const tsconfigOk = FsHelper.existsSync(this.tsconfigPath);
+    const tslintOk = !this.tslintPath || FsHelper.existsSync(this.tslintPath);
 
     // validate logger
     if (this.logger) {
@@ -247,13 +257,13 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  computeContextPath(filePath: string) {
+  private computeContextPath(filePath: string) {
     return path.isAbsolute(filePath)
       ? filePath
       : path.resolve(this.compiler.options.context, filePath);
   }
 
-  pluginStart() {
+  private pluginStart() {
     const run = (_compiler: webpack.Compiler, callback: () => void) => {
       this.isWatching = false;
       callback();
@@ -275,7 +285,7 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  pluginStop() {
+  private pluginStop() {
     const watchClose = () => {
       this.killService();
     };
@@ -301,7 +311,7 @@ class ForkTsCheckerWebpackPlugin {
     });
   }
 
-  registerCustomHooks() {
+  private registerCustomHooks() {
     if (
       this.compiler.hooks.forkTsCheckerServiceBeforeStart ||
       this.compiler.hooks.forkTsCheckerCancel ||
@@ -371,7 +381,7 @@ class ForkTsCheckerWebpackPlugin {
     });
   }
 
-  pluginCompile() {
+  private pluginCompile() {
     if ('hooks' in this.compiler) {
       // webpack 4
       this.compiler.hooks.compile.tap(checkerPluginName, () => {
@@ -389,13 +399,13 @@ class ForkTsCheckerWebpackPlugin {
           this.started = process.hrtime();
 
           // create new token for current job
-          this.cancellationToken = new CancellationToken(undefined, undefined);
+          this.cancellationToken = new CancellationToken();
           if (!this.service || !this.service.connected) {
             this.spawnService();
           }
 
           try {
-            this.service.send(this.cancellationToken);
+            this.service!.send(this.cancellationToken);
           } catch (error) {
             if (!this.silent && this.logger) {
               this.logger.error(
@@ -439,7 +449,7 @@ class ForkTsCheckerWebpackPlugin {
             }
 
             try {
-              this.service.send(this.cancellationToken);
+              this.service!.send(this.cancellationToken);
             } catch (error) {
               if (!this.silent && this.logger) {
                 this.logger.error(
@@ -461,7 +471,7 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  pluginEmit() {
+  private pluginEmit() {
     const emit = (compilation: any, callback: () => void) => {
       if (this.isWatching && this.async) {
         callback();
@@ -486,7 +496,7 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  pluginDone() {
+  private pluginDone() {
     if ('hooks' in this.compiler) {
       // webpack 4
       this.compiler.hooks.done.tap(
@@ -546,7 +556,7 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  spawnService() {
+  private spawnService() {
     this.service = childProcess.fork(
       path.resolve(
         __dirname,
@@ -562,6 +572,7 @@ class ForkTsCheckerWebpackPlugin {
           TSCONFIG: this.tsconfigPath,
           COMPILER_OPTIONS: JSON.stringify(this.compilerOptions),
           TSLINT: this.tslintPath || '',
+          TSLINTAUTOFIX: this.tslintAutoFix,
           WATCH: this.isWatching ? this.watchPaths.join('|') : '',
           WORK_DIVISION: Math.max(1, this.workersNumber),
           MEMORY_LIMIT: this.memoryLimit,
@@ -628,24 +639,25 @@ class ForkTsCheckerWebpackPlugin {
     );
   }
 
-  killService() {
-    if (this.service) {
-      try {
-        if (this.cancellationToken) {
-          this.cancellationToken.cleanupCancellation();
-        }
+  private killService() {
+    if (!this.service) {
+      return;
+    }
+    try {
+      if (this.cancellationToken) {
+        this.cancellationToken.cleanupCancellation();
+      }
 
-        this.service.kill();
-        this.service = undefined;
-      } catch (e) {
-        if (this.logger && !this.silent) {
-          this.logger.error(e);
-        }
+      this.service.kill();
+      this.service = undefined;
+    } catch (e) {
+      if (this.logger && !this.silent) {
+        this.logger.error(e);
       }
     }
   }
 
-  handleServiceMessage(message: Message) {
+  private handleServiceMessage(message: Message): void {
     if (this.cancellationToken) {
       this.cancellationToken.cleanupCancellation();
       // job is done - nothing to cancel
@@ -662,16 +674,36 @@ class ForkTsCheckerWebpackPlugin {
     if (this.ignoreDiagnostics.length) {
       this.diagnostics = this.diagnostics.filter(
         diagnostic =>
-          this.ignoreDiagnostics.indexOf(
-            parseInt(diagnostic.getCode() as string, 10)
-          ) === -1
+          !this.ignoreDiagnostics.includes(
+            parseInt(diagnostic.code as string, 10)
+          )
       );
     }
 
     if (this.ignoreLints.length) {
       this.lints = this.lints.filter(
-        lint => this.ignoreLints.indexOf(lint.getCode() as string) === -1
+        lint => !this.ignoreLints.includes(lint.code as string)
       );
+    }
+
+    if (this.reportFiles.length) {
+      const reportFilesPredicate = (diagnostic: NormalizedMessage): boolean => {
+        if (diagnostic.file) {
+          const relativeFileName = path.relative(
+            this.compiler.options.context,
+            diagnostic.file
+          );
+          const matchResult = micromatch([relativeFileName], this.reportFiles);
+
+          if (matchResult.length === 0) {
+            return false;
+          }
+        }
+        return true;
+      };
+
+      this.diagnostics = this.diagnostics.filter(reportFilesPredicate);
+      this.lints = this.lints.filter(reportFilesPredicate);
     }
 
     if ('hooks' in this.compiler) {
@@ -694,31 +726,38 @@ class ForkTsCheckerWebpackPlugin {
     }
   }
 
-  handleServiceExit(_code: string | number, signal: string) {
-    if (signal === 'SIGABRT') {
-      // probably out of memory :/
-      if (this.compiler) {
-        if ('hooks' in this.compiler) {
-          // webpack 4
-          this.compiler.hooks.forkTsCheckerServiceOutOfMemory.call();
-        } else {
-          // webpack 2 / 3
-          this.compiler.applyPlugins('fork-ts-checker-service-out-of-memory');
-        }
+  private handleServiceExit(_code: string | number, signal: string) {
+    if (signal !== 'SIGABRT') {
+      return;
+    }
+    // probably out of memory :/
+    if (this.compiler) {
+      if ('hooks' in this.compiler) {
+        // webpack 4
+        this.compiler.hooks.forkTsCheckerServiceOutOfMemory.call();
+      } else {
+        // webpack 2 / 3
+        this.compiler.applyPlugins('fork-ts-checker-service-out-of-memory');
       }
-      if (!this.silent && this.logger) {
-        this.logger.error(
-          this.colors.red(
-            'Type checking and linting aborted - probably out of memory. ' +
-              'Check `memoryLimit` option in ForkTsCheckerWebpackPlugin configuration.'
-          )
-        );
-      }
+    }
+    if (!this.silent && this.logger) {
+      this.logger.error(
+        this.colors.red(
+          'Type checking and linting aborted - probably out of memory. ' +
+            'Check `memoryLimit` option in ForkTsCheckerWebpackPlugin configuration.'
+        )
+      );
     }
   }
 
-  createEmitCallback(compilation: any, callback: () => void) {
+  private createEmitCallback(
+    compilation: webpack.compilation.Compilation,
+    callback: () => void
+  ) {
     return function emitCallback(this: ForkTsCheckerWebpackPlugin) {
+      if (!this.elapsed) {
+        throw new Error('Execution order error');
+      }
       const elapsed = Math.round(this.elapsed[0] * 1e9 + this.elapsed[1]);
 
       if ('hooks' in this.compiler) {
@@ -742,17 +781,17 @@ class ForkTsCheckerWebpackPlugin {
         // webpack message format
         const formatted = {
           rawMessage:
-            message.getSeverity().toUpperCase() +
+            message.severity.toUpperCase() +
             ' ' +
             message.getFormattedCode() +
             ': ' +
-            message.getContent(),
+            message.content,
           message: this.formatter(message, this.useColors),
           location: {
-            line: message.getLine(),
-            character: message.getCharacter()
+            line: message.line,
+            character: message.character
           },
-          file: message.getFile()
+          file: message.file
         };
 
         if (message.isWarningSeverity()) {
@@ -766,13 +805,16 @@ class ForkTsCheckerWebpackPlugin {
     };
   }
 
-  createNoopEmitCallback() {
+  private createNoopEmitCallback() {
     // tslint:disable-next-line:no-empty
     return function noopEmitCallback() {};
   }
 
-  createDoneCallback() {
+  private createDoneCallback() {
     return function doneCallback(this: ForkTsCheckerWebpackPlugin) {
+      if (!this.elapsed) {
+        throw new Error('Execution order error');
+      }
       const elapsed = Math.round(this.elapsed[0] * 1e9 + this.elapsed[1]);
 
       if (this.compiler) {