+++ test/integration/index.spec.js
@@ -1,3 +1,4 @@
+var fs = require('fs');
 var describe = require('mocha').describe;
 var it = require('mocha').it;
 var chai = require('chai');
@@ -9,12 +10,31 @@ chai.config.truncateThreshold = 0;
 var expect = chai.expect;
 
 var webpackMajorVersion = require('./webpackVersion')();
+const writeContentsToLintingErrorFile = (fileName, data) => {
+  const promise = new Promise((resolve, reject) => {
+    try {
+      fs.writeFileSync(
+        path.resolve(__dirname, `./project/src/${fileName}.ts`),
+        data,
+        { flag: 'w' }
+      );
+    } catch (e) {
+      return reject(e);
+    }
+    return resolve();
+  });
+  return promise;
+};
 
 describe('[INTEGRATION] index', function() {
   this.timeout(60000);
   var plugin;
 
-  function createCompiler(options, happyPackMode) {
+  function createCompiler(
+    options,
+    happyPackMode,
+    entryPoint = './src/index.ts'
+  ) {
     plugin = new ForkTsCheckerWebpackPlugin(
       Object.assign({}, options, { silent: true })
     );
@@ -23,24 +43,26 @@ describe('[INTEGRATION] index', function() {
       ? { happyPackMode: true, silent: true }
       : { transpileOnly: true, silent: true };
 
-    return webpack({
-      ...(webpackMajorVersion >= 4 ? { mode: 'development' } : {}),
-      context: path.resolve(__dirname, './project'),
-      entry: './src/index.ts',
-      output: {
-        path: path.resolve(__dirname, '../../tmp')
-      },
-      module: {
-        rules: [
-          {
-            test: /\.tsx?$/,
-            loader: 'ts-loader',
-            options: tsLoaderOptions
-          }
-        ]
-      },
-      plugins: [plugin]
-    });
+    return webpack(Object.assign(
+      webpackMajorVersion >= 4 ? { mode: 'development' } : {},
+      {
+        context: path.resolve(__dirname, './project'),
+        entry: entryPoint,
+        output: {
+            path: path.resolve(__dirname, '../../tmp')
+        },
+        module: {
+            rules: [
+            {
+                test: /\.tsx?$/,
+                loader: 'ts-loader',
+                options: tsLoaderOptions
+            }
+            ]
+        },
+        plugins: [plugin]
+      }
+    ));
   }
 
   /**
@@ -55,32 +77,32 @@ describe('[INTEGRATION] index', function() {
   it('should allow to pass no options', function() {
     expect(function() {
       new ForkTsCheckerWebpackPlugin();
-    }).to.not.throw.error;
+    }).to.not.throw();
   });
 
   it('should detect paths', function() {
     var plugin = new ForkTsCheckerWebpackPlugin({ tslint: true });
 
-    expect(plugin.tsconfig).to.be.equal('./tsconfig.json');
-    expect(plugin.tslint).to.be.equal('./tslint.json');
+    expect(plugin.tsconfig).to.equal('./tsconfig.json');
+    expect(plugin.tslint).to.equal('./tslint.json');
   });
 
   it('should set logger to console by default', function() {
     var plugin = new ForkTsCheckerWebpackPlugin({});
 
-    expect(plugin.logger).to.be.equal(console);
+    expect(plugin.logger).to.equal(console);
   });
 
   it('should set watch to empty array by default', function() {
     var plugin = new ForkTsCheckerWebpackPlugin({});
 
-    expect(plugin.watch).to.be.deep.equal([]);
+    expect(plugin.watch).to.deep.equal([]);
   });
 
   it('should set watch to one element array for string', function() {
     var plugin = new ForkTsCheckerWebpackPlugin({ watch: '/test' });
 
-    expect(plugin.watch).to.be.deep.equal(['/test']);
+    expect(plugin.watch).to.deep.equal(['/test']);
   });
 
   it('should work without configuration', function(callback) {
@@ -90,6 +112,95 @@ describe('[INTEGRATION] index', function() {
       expect(stats.compilation.errors.length).to.be.at.least(1);
       callback();
     });
+  });
+
+  it('should fix linting errors with tslintAutofix flag set to true', function(callback) {
+    const lintErrorFileContents = `function someFunctionName(param1,param2){return param1+param2};
+`;
+    const formattedFileContents = `function someFunctionName(param1, param2) {return param1 + param2; }
+`;
+    const fileName = 'lintingError1';
+    writeContentsToLintingErrorFile(fileName, lintErrorFileContents).then(
+      () => {
+        var compiler = createCompiler(
+          {
+            tslintAutoFix: true,
+            tslint: path.resolve(__dirname, './project/tslint.autofix.json'),
+            tsconfig: false
+          },
+          false,
+          `./src/${fileName}.ts`
+        );
+        const deleteFile = () =>
+          fs.unlinkSync(
+            path.resolve(__dirname, `./project/src/${fileName}.ts`)
+          );
+        compiler.run(function(err, stats) {
+          expect(stats.compilation.warnings.length).to.be.eq(0);
+          let fileContents;
+          try {
+            fileContents = fs.readFileSync(
+              path.resolve(__dirname, `./project/src/${fileName}.ts`),
+              {
+                encoding: 'utf-8'
+              }
+            );
+          } catch (e) {
+            throw e;
+          }
+          /*
+            Helpful to wrap this in a try catch.
+            If the assertion fails we still need to cleanup
+            the temporary file created as part of the test
+          */
+          try {
+            expect(fileContents).to.be.eq(formattedFileContents);
+          } catch (e) {
+            deleteFile();
+            throw e;
+          }
+          deleteFile();
+          callback();
+        });
+      },
+      err => {
+        throw err;
+      }
+    );
+  });
+  it('should not fix linting by default', function(callback) {
+    const lintErrorFileContents = `function someFunctionName(param1,param2){return param1+param2};
+`;
+    const fileName = 'lintingError2';
+    const deleteFile = () =>
+      fs.unlinkSync(path.resolve(__dirname, `./project/src/${fileName}.ts`));
+    writeContentsToLintingErrorFile(fileName, lintErrorFileContents).then(
+      () => {
+        var compiler = createCompiler(
+          { tslint: true },
+          false,
+          `./src/${fileName}.ts`
+        );
+        compiler.run(function(err, stats) {
+          /*
+            Helpful to wrap this in a try catch.
+            If the assertion fails we still need to cleanup
+            the temporary file created as part of the test
+          */
+          try {
+            expect(stats.compilation.warnings.length).to.be.eq(7);
+          } catch (e) {
+            deleteFile();
+            throw e;
+          }
+          deleteFile();
+          callback();
+        });
+      },
+      err => {
+        throw err;
+      }
+    );
   });
 
   it('should block emit on build mode', function(callback) {
@@ -190,7 +301,7 @@ describe('[INTEGRATION] index', function() {
       createCompiler({
         tsconfig: '/some/path/that/not/exists/tsconfig.json'
       });
-    }).to.throw.error;
+    }).to.throw();
   });
 
   it('should throw error if config container wrong tslint.json path', function() {
@@ -198,7 +309,7 @@ describe('[INTEGRATION] index', function() {
       createCompiler({
         tslint: '/some/path/that/not/exists/tslint.json'
       });
-    }).to.throw.error;
+    }).to.throw();
   });
 
   it('should find the same errors on multi-process mode', function(callback) {
@@ -227,8 +338,8 @@ describe('[INTEGRATION] index', function() {
     });
 
     function compareResults() {
-      expect(errorsA).to.be.deep.equal(errorsB);
-      expect(warningsA).to.be.deep.equal(warningsB);
+      expect(errorsA).to.deep.equal(errorsB);
+      expect(warningsA).to.deep.equal(warningsB);
       callback();
     }
   });
@@ -236,7 +347,7 @@ describe('[INTEGRATION] index', function() {
   it('should detect tslint path for true option', function() {
     expect(function() {
       createCompiler({ tslint: true });
-    }).to.not.throw.error;
+    }).to.not.throw();
   });
 
   it('should allow delaying service-start', function(callback) {
@@ -284,7 +395,7 @@ describe('[INTEGRATION] index', function() {
     var compiler = createCompiler({}, true);
 
     compiler.run(function(error, stats) {
-      expect(stats.compilation.errors.length).to.be.equal(1);
+      expect(stats.compilation.errors.length).to.equal(1);
       callback();
     });
   });
@@ -293,7 +404,20 @@ describe('[INTEGRATION] index', function() {
     var compiler = createCompiler({ checkSyntacticErrors: true }, true);
 
     compiler.run(function(error, stats) {
-      expect(stats.compilation.errors.length).to.be.equal(2);
+      expect(stats.compilation.errors.length).to.equal(2);
+      callback();
+    });
+  });
+
+  it('should only show errors matching paths specified in reportFiles when provided', function(callback) {
+    var compiler = createCompiler(
+      { checkSyntacticErrors: true, reportFiles: ['**/index.ts'] },
+      true
+    );
+
+    compiler.run(function(error, stats) {
+      expect(stats.compilation.errors.length).to.equal(1);
+      expect(stats.compilation.errors[0].file.endsWith('index.ts')).to.be.true;
       callback();
     });
   });